% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clovers.R
\docType{package}
\name{clovers}
\alias{clovers}
\title{clovers: Probabilistic Models of Customer Lifetime Value (LTV).}
\description{
Analyze customer retention and lifetime value (LTV)
for contractual, discrete-time businesses using Fader and
Hardie's Shifted Beta-Geometric (sBG) probability model.
The package provides functions for data preparation, parameter
estimation, model checking, LTV estimation, and plotting.
}
\section{References}{

Fader, P.S. and Hardie, B.G.S. (2007), How to project customer retention.
\emph{J. Interactive Mark.}, 21: 76-90. doi:10.1002/dir.20074
}

