% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{customer_lifetimes}
\alias{customer_lifetimes}
\title{Individual-level customer lifetimes}
\format{A data frame with 1000 rows and 2 variables:
\describe{
  \item{t}{number of periods}
  \item{is_active}{flag indicating whether the user is still
                    active (1, else 0 if churned)}
}}
\usage{
customer_lifetimes
}
\description{
A dataset of simulated customer lifetimes.
}
\keyword{datasets}
