% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_sbg}
\alias{fit_sbg}
\title{Fit Fader's Shifted Beta Geometric (sBG) model}
\usage{
fit_sbg(data, init_params = c(1, 10))
}
\arguments{
\item{data}{Data frame with the following columns:
\itemize{
    \item \code{t}: integer, the period number (\code{t} = 1,2,3,...).
    \item \code{n_churn}: integer, the number of customers who
        churned after \code{t} periods.
    \item \code{n_survive}: integer, the number of customers who
        have been subscribed for exactly \code{t} periods and
        are still subscribed.
}}

\item{init_params}{A list containing initial values of the
sBG parameters \code{a} and \code{b}.}
}
\value{
A list with the following elements
    \item{a}{ Numeric, the \code{a} parameter of the sBG model.}
    \item{b}{ Numeric, the \code{b} parameter of the sBG model.}
}
\description{
Fit Fader's Shifted Beta Geometric (sBG) model
}
\examples{
data(customer_lifetimes)
data = transform_individual(customer_lifetimes)
fit_sbg(data)
}
