% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{lifetime_stats}
\alias{lifetime_stats}
\title{Mean and standard deviation of modeled lifetime.}
\usage{
lifetime_stats(fit, horizon = 36)
}
\arguments{
\item{fit}{List with components \code{a} and \code{b},
the parameters of the sBG model}

\item{horizon}{Integer, the number of periods defining the horizon.}
}
\value{
A list with the following elements:
    \item{\code{mean_lifetime}}{Numeric, the mean lifetime predicted
         by the model.}
    \item{\code{sd_lifetime}}{Numeric, the standard deviation of the lifetime
        predicted by the model.}
}
\description{
Compute mean and SD of lifetime, limited to a time horizon,
for a fitted model. Uses the \code{a} and \code{b} parameters of the
beta distribution (as returned by, e.g., the function
\code{fit_sbg}).
}
\details{
The lifetime stats are computed assuming that we limit the data
to a finite number of periods, as specified by the horizon parameter.
Thus one can compute, for example, a 12-month or 36-month LTV.
}
\examples{
data(customer_lifetimes)
data = transform_individual(customer_lifetimes)
fit = fit_sbg(data)
lifetime_stats(fit)
}
