% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{loglik_sbg}
\alias{loglik_sbg}
\title{Log-likelihood function for the sBG model}
\usage{
loglik_sbg(loga, logb, data)
}
\arguments{
\item{loga}{Numeric, the natural log of the \code{a} parameter
of the sBG model.}

\item{logb}{Numeric, the natural log of the \code{b} parameter
of the sBG model.}

\item{data}{Data frame with the following columns:
\itemize{
    \item \code{t}: integer, the period number (\code{t} = 1,2,3,...).
    \item \code{n_churn}: integer, the number of customers who
        churned after \code{t} periods.
    \item \code{n_survive}: integer, the number of customers who
        have been subscribed for exactly \code{t} periods and
        are still subscribed.
}}
}
\value{
The log-likelihood for the sBG model.
}
\description{
Log-likelihood function for the sBG model
}
