% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_churn_distribution}
\alias{plot_churn_distribution}
\title{Plot distribution of churn rates}
\usage{
plot_churn_distribution(fit, type = "density", sample_size = 1000, n_bins = 10)
}
\arguments{
\item{fit}{List with components \code{a} and \code{b},
the parameters of the sBG model}

\item{type}{Type of plot to draw: \code{"density"},
\code{"count"}, or \code{"percent"}.}

\item{sample_size}{Integer. If \code{type="count"}, total number
of customers.}

\item{n_bins}{Integer. If \code{type="count"} or \code{type="percent"},
number of bins.}
}
\value{
A list containing the following components:
    \item{plot}{A ggplot2 object containing the plot.}
    \item{data}{A data frame containing the plot data.}
}
\description{
Plots the distribution of churn rates found by the sBG model.
}
\details{
With \code{type="density"}, the probability density function is plotted.

With \code{type="count"}, a set of customers numbering \code{sample_size}
is assumed, the churn rates are binned into \code{n_bins} bins, and the
count of customers in each bin is plotted as a bar chart.

With \code{type="percent"}, the plot is similar to the plot produced by
\code{type="count"}, except that the bar heights represent the percentages
of customers in each bin instead of the counts.
}
\examples{
data(customer_lifetimes)
data = transform_individual(customer_lifetimes)
fit = fit_sbg(data)
plot_churn_distribution(fit)
plot_churn_distribution(list(a=1.5, b=1.5), type="count")
}
