% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_model_fit}
\alias{plot_model_fit}
\title{Plot empirical survival and modeled survival}
\usage{
plot_model_fit(fit, data)
}
\arguments{
\item{fit}{List with components \code{a} and \code{b},
the parameters of the sBG model}

\item{data}{Data frame with the following columns:
\itemize{
    \item \code{t}: integer, the period number (\code{t} = 1,2,3,...).
    \item \code{n_churn}: integer, the number of customers who
        churned after \code{t} periods.
    \item \code{n_survive}: integer, the number of customers who
        have been subscribed for exactly \code{t} periods and
        are still subscribed.
}}
}
\value{
A list containing the following components:
    \item{plot}{A ggplot2 object containing the plot.}
    \item{data}{A data frame containing the plot data.}
}
\description{
Draws a plot with two curves: the empirical survival
function computed with the Kaplan-Meier estimator, and
the survival function computed from the fitted sBG model.
}
\examples{
data(customer_lifetimes)
data = transform_individual(customer_lifetimes)
fit = fit_sbg(data)
plot_model_fit(fit, data)
}
