% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_survival_estimate}
\alias{plot_survival_estimate}
\title{Plot empirical survival function}
\usage{
plot_survival_estimate(data)
}
\arguments{
\item{data}{Data frame with the following columns:
\itemize{
    \item \code{t}: integer, the period number (\code{t} = 1,2,3,...).
    \item \code{n_churn}: integer, the number of customers who
        churned after \code{t} periods.
    \item \code{n_survive}: integer, the number of customers who
        have been subscribed for exactly \code{t} periods and
        are still subscribed.
}}
}
\value{
A list containing the following components:
    \item{plot}{A ggplot2 object containing the plot.}
    \item{data}{A data frame containing the plot data.}
}
\description{
Plots the Kaplan-Meier estimate of the survival function.
}
\examples{
data(customer_lifetimes)
data = transform_individual(customer_lifetimes)
plot_survival_estimate(data)
}
