% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{survival_estimate}
\alias{survival_estimate}
\title{Compute empirical survival function}
\usage{
survival_estimate(data)
}
\arguments{
\item{data}{Data frame with the following columns:
\itemize{
    \item \code{t}: integer, the period number (\code{t} = 1,2,3,...).
    \item \code{n_churn}: integer, the number of customers who
        churned after \code{t} periods.
    \item \code{n_survive}: integer, the number of customers who
        have been subscribed for exactly \code{t} periods and
        are still subscribed.
}}
}
\value{
A data frame with columns
\itemize{
   \item \code{t}: Integer, the number of periods.
   \item \code{frac_surv}: Numeric, the fraction of customers still active
       after \code{t} periods.
 }
}
\description{
Computes the Kaplan-Meier estimate of the survival function.
}
\examples{
data(customer_lifetimes)
data = transform_individual(customer_lifetimes)
survival_estimate(data)
}
