% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_individual}
\alias{transform_individual}
\title{Transform individual-level data}
\usage{
transform_individual(individual)
}
\arguments{
\item{individual}{A data frame with one row per individual, and columns
\itemize{
  \item \code{t}: Integer, the number of periods for which the customer
      has been active.
  \item \code{is_active}: Integer, a flag indicating whether the customer
      is still active (\code{1} if still active, \code{0} if churned).
}}
}
\value{
Data frame with the following columns:
    \itemize{
        \item \code{t}: integer, the period number (\code{t} = 1,2,3,...).
        \item \code{n_churn}: integer, the number of customers who
            churned after \code{t} periods.
        \item \code{n_survive}: integer, the number of customers who
            have been subscribed for exactly \code{t} periods and
            are still subscribed.
    }
}
\description{
Aggregate individual-level data into aggregate format
required by the fit_sbg function.
}
\examples{
data(customer_lifetimes)
transform_individual(customer_lifetimes)
}
